+++ packages/date-time-utilities/.npmignore
@@ -1,0 +1,32 @@
+*.api.json
+*.config.js
+*.log
+*.nuspec
+*.test.*
+*.yml
+.editorconfig
+.gitattributes
+.gitignore
+.vscode
+coverage
+dist-storybook
+dist/*.stats.html
+dist/*.stats.json
+dist/demo*.*
+fabric-test*
+gulpfile.js
+images
+index.html
+jsconfig.json
+node_modules
+results
+src/**/*
+!src/**/examples/*.tsx
+!src/**/docs/**/*.md
+!src/**/*.types.ts
+temp
+tsconfig.json
+tsd.json
+tslint.json
+typings
+visualtests
