+++ packages/date-time-utilities/README.md
@@ -1,0 +1,11 @@
+# @fluentui/date-time-utilities
+
+**Date Time Utilities components for [Fluent UI React](https://developer.microsoft.com/en-us/fluentui)**
+
+These are not production-ready components and **should never be used in product**. This space is useful for testing new components whose APIs might change before final release.
+
+To import Date Time Utilities components:
+
+```js
+import { ComponentName } from '@fluentui/date-time-utilities';
+```
