+++ packages/date-time-utilities/etc/date-time-utilities.api.md
@@ -1,0 +1,126 @@
+## API Report File for "@uifabric/date-time-utilities"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+// @public
+export function addDays(date: Date, days: number): Date;
+
+// @public
+export function addMonths(date: Date, months: number): Date;
+
+// @public
+export function addWeeks(date: Date, weeks: number): Date;
+
+// @public
+export function addYears(date: Date, years: number): Date;
+
+// @public
+export function compareDatePart(date1: Date, date2: Date): Number;
+
+// @public
+export function compareDates(date1: Date, date2: Date): boolean;
+
+// @public
+export enum DateRangeType {
+    // (undocumented)
+    Day = 0,
+    // (undocumented)
+    Month = 2,
+    // (undocumented)
+    Week = 1,
+    // (undocumented)
+    WorkWeek = 3
+}
+
+// @public
+export enum DayOfWeek {
+    // (undocumented)
+    Friday = 5,
+    // (undocumented)
+    Monday = 1,
+    // (undocumented)
+    Saturday = 6,
+    // (undocumented)
+    Sunday = 0,
+    // (undocumented)
+    Thursday = 4,
+    // (undocumented)
+    Tuesday = 2,
+    // (undocumented)
+    Wednesday = 3
+}
+
+// @public
+export enum FirstWeekOfYear {
+    // (undocumented)
+    FirstDay = 0,
+    // (undocumented)
+    FirstFourDayWeek = 2,
+    // (undocumented)
+    FirstFullWeek = 1
+}
+
+// @public
+export function getDateRangeArray(date: Date, dateRangeType: DateRangeType, firstDayOfWeek: DayOfWeek, workWeekDays?: DayOfWeek[], daysToSelectInDayView?: number): Date[];
+
+// @public
+export function getMonthEnd(date: Date): Date;
+
+// @public
+export function getMonthStart(date: Date): Date;
+
+// @public
+export function getStartDateOfWeek(date: Date, firstDayOfWeek: DayOfWeek): Date;
+
+// @public
+export function getWeekNumber(date: Date, firstDayOfWeek: DayOfWeek, firstWeekOfYear: FirstWeekOfYear): number;
+
+// @public
+export function getWeekNumbersInMonth(weeksInMonth: number, firstDayOfWeek: DayOfWeek, firstWeekOfYear: FirstWeekOfYear, navigatedDate: Date): number[];
+
+// @public
+export function getYearEnd(date: Date): Date;
+
+// @public
+export function getYearStart(date: Date): Date;
+
+// @public
+export function isInDateRangeArray(date: Date, dateRange: Date[]): boolean;
+
+// @public
+export enum MonthOfYear {
+    // (undocumented)
+    April = 3,
+    // (undocumented)
+    August = 7,
+    // (undocumented)
+    December = 11,
+    // (undocumented)
+    February = 1,
+    // (undocumented)
+    January = 0,
+    // (undocumented)
+    July = 6,
+    // (undocumented)
+    June = 5,
+    // (undocumented)
+    March = 2,
+    // (undocumented)
+    May = 4,
+    // (undocumented)
+    November = 10,
+    // (undocumented)
+    October = 9,
+    // (undocumented)
+    September = 8
+}
+
+// @public
+export function setMonth(date: Date, month: number): Date;
+
+
+// (No @packageDocumentation comment for this package)
+
+```
