+++ packages/date-time-utilities/just.config.ts
@@ -1,0 +1,4 @@
+const { preset, just } = require('@uifabric/build');
+const { chain } = just;
+
+preset();
