+++ packages/date-time-utilities/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@uifabric/date-time",
-  "version": "7.8.100",
-  "description": "Date and time related React components for building experiences for Microsoft 365.",
+  "name": "@uifabric/date-time-utilities",
+  "version": "0.1.0",
+  "description": "Date-time utilities",
   "main": "lib-commonjs/index.js",
   "module": "lib/index.js",
   "typings": "lib/index.d.ts",
@@ -16,45 +16,32 @@
   "scripts": {
     "build": "just-scripts build",
     "bundle": "just-scripts bundle",
-    "lint": "just-scripts lint",
     "test": "just-scripts test",
     "just": "just-scripts",
-    "clean": "just-scripts clean",
     "code-style": "just-scripts code-style",
-    "start": "just-scripts dev:storybook",
+    "clean": "just-scripts clean",
     "start-test": "just-scripts jest-watch",
-    "update-snapshots": "just-scripts jest -u",
+    "lint": "just-scripts lint",
     "update-api": "just-scripts update-api"
   },
   "devDependencies": {
     "@types/enzyme": "3.10.3",
     "@types/enzyme-adapter-react-16": "1.0.3",
     "@types/es6-promise": "0.0.32",
-    "@types/jest": "~24.9.0",
-    "@types/prop-types": "15.7.1",
+    "@types/jest": "^24.9.1",
     "@types/react": "16.8.25",
     "@types/react-dom": "16.8.4",
-    "@types/react-test-renderer": "^16.0.0",
     "@types/webpack-env": "1.15.1",
     "@uifabric/build": "^7.0.0",
-    "@uifabric/example-app-base": "^7.12.79",
-    "@uifabric/jest-serializer-merge-styles": "^7.0.26",
     "@uifabric/tslint-rules": "^7.2.1",
-    "enzyme": "~3.10.0",
-    "enzyme-adapter-react-16": "^1.15.0",
     "es6-weak-map": "^2.0.2",
     "react": "16.8.6",
     "react-app-polyfill": "~1.0.1",
-    "react-dom": "16.8.6",
-    "react-test-renderer": "^16.3.0"
+    "react-dom": "16.8.6"
   },
   "dependencies": {
-    "@microsoft/load-themed-styles": "^1.10.26",
-    "@uifabric/date-time-utilities": "^0.1.0",
     "@uifabric/set-version": "^7.0.13",
-    "@uifabric/styling": "^7.12.18",
-    "@uifabric/utilities": "^7.21.0",
-    "office-ui-fabric-react": "^7.120.1",
+    "@uifabric/utilities": "^7.20.3",
     "tslib": "^1.10.0"
   },
   "peerDependencies": {
