+++ packages/date-time-utilities/tslint.json
@@ -1,0 +1,4 @@
+{
+  "extends": ["@uifabric/tslint-rules"],
+  "rules": {}
+}
