+++ packages/date-time/etc/date-time.api.md
@@ -387,8 +387,8 @@ export interface IWeeklyDayPickerStyles extends Partial<ICalendarDayGridStyles> 
 export const WeeklyDayPicker: React.FunctionComponent<IWeeklyDayPickerProps>;
 
 
-export * from "office-ui-fabric-react/lib/utilities/dateMath/DateMath";
-export * from "office-ui-fabric-react/lib/utilities/dateValues/DateValues";
+export * from "@uifabric/date-time-utilities/lib/dateMath/DateMath";
+export * from "@uifabric/date-time-utilities/lib/dateValues/DateValues";
 
 // (No @packageDocumentation comment for this package)
