+++ packages/office-ui-fabric-react/jest.config.js
@@ -8,6 +8,7 @@ const config = createConfig({
     '@fluentui/react-focus/lib/(.*)$': '@fluentui/react-focus/lib-commonjs/$1',
     '@uifabric/react-hooks/lib/(.*)$': '@uifabric/react-hooks/lib-commonjs/$1',
     '@uifabric/utilities/lib/(.*)$': '@uifabric/utilities/lib-commonjs/$1',
+    '@uifabric/date-time-utilities/lib/(.*)$': '@uifabric/date-time-utilities/lib-commonjs/$1',
     // These mappings allow Jest to run snapshot tests against Example files.
     'office-ui-fabric-react/lib/(.*)$': '<rootDir>/src/$1',
     'office-ui-fabric-react$': '<rootDir>/src/',
