+++ packages/office-ui-fabric-react/src/utilities/dateValues/TimeConstants.ts
@@ -1,15 +1,1 @@
-/**
- * Common constants and methods
- */
-const TimeConstants = {
-  MillisecondsInOneDay: 86400000,
-  MillisecondsIn1Sec: 1000,
-  MillisecondsIn1Min: 60000,
-  MillisecondsIn30Mins: 1800000,
-  MillisecondsIn1Hour: 3600000,
-  MinutesInOneDay: 1440,
-  MinutesInOneHour: 60,
-  DaysInOneWeek: 7,
-  MonthInOneYear: 12,
-};
-export default TimeConstants;
+export * from '@uifabric/date-time-utilities/lib/dateValues/TimeConstants';
